Prototype action to build level 3 tower Add grid display to MotionController testing

+++ miam_utils/src/MathUtils.cpp
@@ -8,9 +8,9 @@ std::vector<double> movingAverage(std::vector<double> const values, unsigned int
     std::vector<double> outputValue;
     for (unsigned int i = 0; i < values.size(); i++)
     {
-        unsigned int start = i - winSize;
-        unsigned int end = i + winSize + 1;
-        if (end > values.size())
+        int start = i - winSize;
+        int end = i + winSize + 1;
+        if (end > static_cast<int>(values.size()))
         {
             int shift = end - values.size();
             end -= shift;
